pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nomad-oasis-test-plugin"
version = "0.1.0"
description = "Minimal test plugin for NOMAD Oasis"
authors = [{name = "Gemma"}]
requires-python = ">=3.10"
dependencies = ["nomad-lab"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."nomad.plugin"]
test_schema = "nomad_oasis_test_plugin.schema_packages.plugin:TestSchema"
